Feat: Added security, nav-menu and 404 page

+++ php/likepost.php
@@ -1,4 +1,5 @@
 <?php 
+require_once('SqlConnector.php');
 // get user id 
 session_start();
 if (isset($_SESSION['user_id'])) {
@@ -17,30 +18,35 @@ else {
 $post_id = $_POST['post_id'];
 
 // connect to the database
+$sqlConnector = new SqlConnector();
 
-$servername = "localhost";
-$username = "root";
-$password = "root";
-$dbname = "we4a_blog_db";
-
-// Create connection
-$conn = new mysqli($servername, $username, $password, $dbname);
+if ($sqlConnector->is_working == false) {
+    die("Connection failed: " . $sqlConnector->conn->connect_error);
+}
 
 // Sql query to check if the user has already liked the post
 $sql = "SELECT * FROM post_likes WHERE user_id = '".$user_id."' AND post_id = '".$post_id."'";
 
-$result = $conn->query($sql);
+$result = $sqlConnector->ask_database($sql);
 
 // check if the user has already liked the post
-if ($result->num_rows > 0) {
-    //Already liked
+if ($result != null && $result->num_rows > 0) {
+    //Already liked so we delete the like
     $sql2 = "DELETE FROM post_likes WHERE user_id = '".$user_id."' AND post_id = '".$post_id."'";
-    $result2 = $conn->query($sql2);
+    $result2 = $sqlConnector->ask_database($sql2);
+
+    // Decrease the number of likes of the post
+    $sql3 = "UPDATE userpost SET likes = likes - 1 WHERE post_id = '".$post_id."'";
+    $result3 = $sqlConnector->ask_database($sql3);
 } 
 else {
-    //Not already liked
+    //Not already liked so we add the like
     $sql2 = "INSERT INTO post_likes (user_id, post_id) VALUES ('".$user_id."', '".$post_id."')";
-    $result2 = $conn->query($sql2);
+    $result2 = $sqlConnector->ask_database($sql2);
+
+    // Increase the number of likes of the post
+    $sql3 = "UPDATE userpost SET likes = likes + 1 WHERE post_id = '".$post_id."'";
+    $result3 = $sqlConnector->ask_database($sql3);
 }
 
 $response = array( 
@@ -48,7 +54,4 @@ $response = array(
 );
 header("Content-Type: application/json");
 echo json_encode($response);
-
-// close the connection
-$conn->close();
 ?>
